hijacks-2/pythonCode/requirements.txt
git+https://github.com/ANSSI-FR/tabi#egg=tabi
kafka-python
prometheus_client
